lint/refactor handleEntityChanged
