+++ example/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>My Scene</title>
-    <script src="js/aframe.js"></script>
+    <script src="js/aframe.min.js"></script>
   </head>
   <body>
     <a-scene debug="true" stats>
